# backend/docs/swagger.yaml
basePath: /
definitions:
...
  utils.ErrorDetail:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  utils.ErrorResponse:
    properties:
      error:
        properties:
          code:
            type: string
          details:
            items:
              $ref: '#/definitions/utils.ErrorDetail'
            type: array
          message:
            type: string
          request_id:
            type: string
          timestamp:
            type: string
        type: object
    type: object
...
host: localhost:8080
info:
  contact:
    name: API Support
    url: https://example.com/support
  description: API for Reelcut — clip creation from long-form video (auth, projects,
    videos, transcriptions, analysis, clips, templates, jobs).
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://example.com/terms
  title: Reelcut API
  version: "1.0"
paths:
...
  /api/v1/projects:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: List projects
      tags:
      - projects
    post:
      consumes:
      - application/json
      parameters:
      - description: name, description
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Create a project
      tags:
      - projects
  /api/v1/projects/{id}:
    delete:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a project
      tags:
      - projects
    get:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get project by ID
      tags:
      - projects
    put:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: name, description
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a project
      tags:
      - projects
  /api/v1/templates:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: List current user's templates
      tags:
      - templates
    post:
      consumes:
      - application/json
      parameters:
      - description: name, category, is_public, style_config
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Create a template
      tags:
      - templates
  /api/v1/templates/{id}:
    delete:
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a template
      tags:
      - templates
    get:
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get template by ID
      tags:
      - templates
    put:
      consumes:
      - application/json
      parameters:
      - description: Template ID
        in: path
        name: id
        required: true
        type: string
      - description: name, category, is_public, style_config
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a template
      tags:
      - templates
  /api/v1/templates/public:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: List public templates (no auth)
      tags:
      - templates
  /api/v1/transcriptions/{id}:
    get:
      parameters:
      - description: Transcription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get transcription by ID
      tags:
      - transcriptions
  /api/v1/transcriptions/{id}/segments/{segmentId}:
    put:
      consumes:
      - application/json
      parameters:
      - description: Transcription ID
        in: path
        name: id
        required: true
        type: string
      - description: Segment ID
        in: path
        name: segmentId
        required: true
        type: string
      - description: text, start_time, end_time
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a transcript segment
      tags:
      - transcriptions
  /api/v1/transcriptions/videos/{videoId}:
    get:
      parameters:
      - description: Video ID
        in: path
        name: videoId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get transcription by video ID
      tags:
      - transcriptions
    post:
      consumes:
      - application/json
      parameters:
      - description: Video ID
        in: path
        name: videoId
        required: true
        type: string
      - description: language, enable_diarization
        in: body
        name: body
        schema:
          type: object
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Start transcription for a video
      tags:
      - transcriptions
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"